refactor: Event tests should be called "events" not "custom events" because it also tests built-in e
